+++ benchmarks/benchmark_indicators.py
@@ -137,6 +137,7 @@ INDICATORS: tuple[IndicatorSpec, ...] = (
     IndicatorSpec("KalmanInnovationZScore", ("close",), batch_inputs=("close",)),
     IndicatorSpec("KalmanLocalLinearTrend", ("close",), batch_inputs=("close",)),
     IndicatorSpec("KalmanMovingAverage", ("close",), batch_inputs=("close",)),
+    IndicatorSpec("KalmanPredictionBands", ("close",), batch_inputs=("close",)),
     IndicatorSpec("KalmanVelocityOscillator", ("close",), batch_inputs=("close",)),
     IndicatorSpec("Kama", ("close",), batch_inputs=("input",)),
     IndicatorSpec("KeltnerChannel", ("close", "high", "low")),